+++ examples/nss-runfile.babel.js/nss-runfile.babel.js
@@ -1,0 +1,3 @@
+import { task } from "../../lib";
+
+task('test', () => console.log('Hello world: nss-runfile.mjs'));