+++ examples/nss-runfile.js/nss-runfile.js
@@ -1,0 +1,3 @@
+const nssRun = require('../../lib/index');
+
+nssRun.task('test', () => console.log('Hello world: nss-runfile.js'));