+++ examples/nss-runfile.mjs/nss-runfile.mjs
@@ -1,0 +1,3 @@
+import { task } from "../../lib/index.js";
+
+task('test', () => console.log('Hello world: nss-runfile.mjs'));